[REFACT] move online registration code to another module
