+++ dc_website_registration/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+# © 2016 Coop IT Easy (http://www.coopiteasy.be)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+from . import controllers
+from . import models
