+++ dc_website_registration/controllers/__init__.py
@@ -1,0 +1,1 @@
+from . import authsignup
