+++ dc_website_registration/controllers/authsignup.py
@@ -1,3 +1,6 @@
+# Copyright 2017 Coop IT Easy SCRLfs
+#     Houssine Bakkali <houssine@coopiteasy.be>
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 import logging
 import werkzeug
