+++ dc_website_registration/models/__init__.py
@@ -1,0 +1,2 @@
+from . import res_users
+from . import partner
