+++ dc_website_registration/static/src/js/auth_signup.js
