+++ dc_website_registration/views/auth_signup_template.xml
@@ -1,6 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 	<data>
+		<template id="assets_frontend" inherit_id="website.assets_frontend" name="Distribution Circtuits Registration">
+			<xpath expr="." position="inside">
+				  <script src="/dc_website_registration/static/src/js/auth_signup.js" type="text/javascript"/>
+			</xpath>
+		</template>
+
 		<template id="easy_my_hub_auth_fields" inherit_id="auth_signup.fields" name="Auth Signup/ResetPassword form fields">
 			<xpath expr="//input[@name='name']/.." position="replace"/>
 			<xpath expr="//input[@name='login']/.." position="after">
