+++ dc_website_registration/views/res_users_view.xml
