+++ distribution_circuits_website_sale/__init__.py
@@ -1,6 +1,2 @@
-# -*- coding: utf-8 -*-
-# © 2016 Coop IT Easy (http://www.coopiteasy.be)
-# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
-
 from . import controllers
 from . import models
