+++ distribution_circuits_website_sale/controllers/__init__.py
@@ -1,2 +1,1 @@
 from . import main
-from . import authsignup
