+++ distribution_circuits_website_sale/models/__init__.py
@@ -1,4 +1,3 @@
 from . import res_config_settings
-from . import res_users
 from . import sale_order
 from . import website
